requirements.txt
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
attrs==23.1.0
backoff==2.2.1
bcrypt==4.1.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.21
click==8.1.7
coloredlogs==15.0.1
Deprecated==1.2.14
fastapi==0.108.0
filelock==3.13.1
flatbuffers==23.5.26
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.40
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-auth==2.25.2
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
httptools==0.6.1
huggingface-hub==0.20.1
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kubernetes==28.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mmh3==4.0.1
monotonic==1.6
mpmath==1.3.0
numpy==1.26.2
oauthlib==3.2.2
onnxruntime==1.16.3
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
overrides==7.4.0
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
posthog==3.1.0
proto-plus==1.23.0
protobuf==4.25.1
pulsar-client==3.3.0
pyarrow==14.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
Pygments==2.17.2
PyPDF2==3.0.1
PyPika==0.48.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.15.2
rsa==4.9
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
starlette==0.32.0.post1
streamlit==1.29.0
sympy==1.12
tenacity==8.2.3
tokenizers==0.15.0
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.18
uvicorn==0.25.0
uvloop==0.19.0
validators==0.22.0
watchdog==3.0.0
watchfiles==0.21.0
websocket-client==1.7.0
websockets==12.0
wrapt==1.16.0
zipp==3.17.0
